+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## 0.1.1
+
+- Integrated new Orwynn.
+- KeyCode format to permissions.
+
 ## 0.1.0
 
 - Initial release!